pyproject.toml
[project]
name = "dsa-judge-server"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Takuya Mizokami", email = "takoyaki65@users.noreply.github.com" }
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "pydantic>=2.7.3",
    "docker>=7.1.0",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
